% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_pivot.R
\name{check_function_output}
\alias{check_function_output}
\title{Check output is length 1}
\usage{
check_function_output(output)
}
\arguments{
\item{output}{Output of the value function}
}
\value{
output
}
\description{
Check output is length 1
}
