% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{simplepivot_mean_sd}
\alias{simplepivot_mean_sd}
\title{Simple Pivot Summary Function: Mean (SD)}
\usage{
simplepivot_mean_sd(num, rounding = 3)
}
\arguments{
\item{num}{A numeric vector}

\item{rounding}{A number, of digits to show}
}
\value{
A string of the form "Mean (SD)"
}
\description{
Simple Pivot Summary Function: Mean (SD)
}
\examples{
simplepivot_mean_sd(c(1,2,3,4,5))
}
