% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{simplepivot_median_IQR}
\alias{simplepivot_median_IQR}
\title{Simple Pivot Summary Function: Median (IQR)}
\usage{
simplepivot_median_IQR(num, rounding = 3)
}
\arguments{
\item{num}{A numeric vector}

\item{rounding}{A number, of digits to show}
}
\value{
A string of the form "Median (IQR)"
}
\description{
Simple Pivot Summary Function: Median (IQR)
}
\examples{
simplepivot_median_IQR(c(1,2,3,4,5))
}
