% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_pivot.R
\name{simple_pivot}
\alias{simple_pivot}
\title{Simple Pivot Data Transformation}
\usage{
simple_pivot(df, row_vars, col_vars, value_variable, value_function)
}
\arguments{
\item{df}{Data frame object to transform}

\item{row_vars}{Character vector naming the row variable(s)}

\item{col_vars}{Character vector naming the column variable(s)}

\item{value_variable}{Length 1 character vector naming the column with the value to summarise in table cells}

\item{value_function}{Function that takes a single vector of values as first (and only required) argument and returns a single summary
value (the base r 'mean' function, for example). Used to summarise the 'value variable' (by row/column group).}
}
\value{
a tibble/data.frame with the pivot operation applied to it
}
\description{
Simple Pivot Data Transformation
}
\examples{
simple_pivot(df = mtcars,
 row_vars = c("gear"),
 col_vars = c("carb"),
 value_variable = "mpg",
 value_function = sum)
}
